+++ src/libs/pulumi-cli.ts
@@ -1,8 +1,8 @@
+import * as os from 'os';
+import * as path from 'path';
 import * as core from '@actions/core';
 import * as io from '@actions/io';
 import * as tc from '@actions/tool-cache';
-import * as os from 'os';
-import * as path from 'path';
 import * as semver from 'semver';
 import * as exec from './exec';
 import { getVersionObject } from './libs/get-version';
@@ -10,6 +10,20 @@ import { getVersionObject } from './libs/get-version';
 export async function isAvailable(): Promise<boolean> {
   const res = await exec.exec(`pulumi`, [], true);
   return res.stderr != '' && !res.success ? false : res.success;
+}
+
+export async function getVersion(): Promise<string | undefined> {
+  const res = await exec.exec('pulumi', ['version'], true);
+
+  // Only check for success and if the [stdout] starts with the version
+  // prefix 'v'. If success is true and the runner version is not the 
+  // latest version then the "warning of newer version" will be in [stderr] field, 
+  // which will trigger an else condition if we also check for [stderr === '']
+  if (res.success && res.stdout.startsWith('v'))
+    // Return version without 'v' prefix
+    return res.stdout.substring(1); 
+  else
+    return undefined;
 }
 
 export async function run(...args: string[]): Promise<void> {
@@ -42,6 +56,22 @@ export async function downloadCli(range: string): Promise<void> {
   }
 
   core.info(`Configured range: ${range}`);
+
+  // Check for version of Pulumi CLI installed on the runner
+  const runnerVersion = await getVersion();
+
+  if (runnerVersion) {
+    // Check if runner version matches
+    if (semver.satisfies(runnerVersion, range)) {
+      // If runner version matches, skip downloading CLI by exiting the function
+      core.info(`Pulumi version ${runnerVersion} is already installed on this machine. Skipping download`);
+      return;
+    } else {
+      core.info(`Pulumi ${runnerVersion} does not satisfy the desired version ${range}. Proceeding to download`);
+    }
+  } else {
+    core.info('Pulumi is not detected in the PATH. Proceeding to download');
+  }
 
   const { version, downloads } = await getVersionObject(range);