+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - chore: add dist files to gitignore and forcebly commit them instead
 - refactor: Improve config
+- feat: Add support for skipping downloading the Pulumi CLI if it is already
+  installed on the runner being used and the runner CLI version is suitable.
+  ([#789](https://github.com/pulumi/actions/issues/789))
 
 --